Extract PersistWindowState out
