+++ src/gui/mod.rs
@@ -1,6 +1,7 @@
 mod draw_quad;
 pub mod draw_state;
 pub mod gfx_ui;
+mod persist_window_state;
 
 pub type ColorFormat = gfx::format::Rgba8;
 pub type DepthFormat = gfx::format::Depth;
